fix(core): address PDF warning review feedback
Relax raster-backed text layer detection so scanned pages with minor vector marks still suppress OCR geometry noise.

Clarify bottom page labels and overlap suppression to reduce false positives, and split the warning helpers to keep the detector maintainable.
